to_complex improvements
