+++ setup.cfg
@@ -16,7 +16,7 @@ mistral.expression.functions =
     regex_search = st2mistral.functions.regex:regex_search
     regex_substring = st2mistral.functions.regex:regex_substring
     st2kv = st2mistral.functions.stackstorm:st2kv_
-    to_complex = st2mistral.functions.complex_type:to_complex
+    to_complex = st2mistral.functions.data:to_complex
     to_human_time_from_seconds = st2mistral.functions.time:to_human_time_from_seconds
     to_json_string = st2mistral.functions.data:to_json_string
     to_yaml_string = st2mistral.functions.data:to_yaml_string
