+++ st2mistral/tests/unit/test_jinja_render_data.py
@@ -21,6 +21,16 @@ from st2mistral.tests.unit import test_function_base as base
 
 class JinjaUtilsDataFunctionTestCase(base.JinjaFunctionTestCase):
 
+    def test_function_to_complex(self):
+        env = self.get_jinja_environment()
+        obj = {'a': 'b', 'c': {'d': 'e', 'f': 1, 'g': True}}
+
+        template = '{{k1 | to_complex}}'
+
+        obj_json_str = env.from_string(template).render({'k1': obj})
+        actual_obj = json.loads(obj_json_str)
+        self.assertDictEqual(obj, actual_obj)
+
     def test_function_to_json_string(self):
         env = self.get_jinja_environment()
         obj = {'a': 'b', 'c': {'d': 'e', 'f': 1, 'g': True}}
